Remove combinators
